requirements.txt
fastapi==0.116.1
python-dotenv==1.1.1
langchain==0.3.27